fix: correct column handling and setting focus
