+++ src/table/components/withSelections.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { useContextSelector, TableContext } from '../context';
 import { getSelectionStyle } from '../utils/styling-utils';
-import { getCellSelectionState, SelectionActions } from '../utils/selections-utils';
+import { getCellSelectionState } from '../utils/selections-utils';
+import { getMouseHandlers } from '../utils/handle-click';
 import { CellHOC, CellHOCProps } from '../types';
 
 export default function withSelections(CellComponent: CellHOC) {
@@ -12,24 +13,12 @@ export default function withSelections(CellComponent: CellHOC) {
     );
     const selectionDispatch = useContextSelector(TableContext, (value) => value.selectionDispatch);
 
-    const handleMouseDown = () => {
-      selectionDispatch({ type: SelectionActions.SELECT_MULTI_START, payload: { cell } });
-    };
-
-    const handleMouseOver = (evt: React.MouseEvent) => {
-      if (evt.button === 0 && evt.buttons === 1)
-        selectionDispatch({ type: SelectionActions.SELECT_MULTI_ADD, payload: { cell, evt, announce } });
-    };
-
-    const handleMouseUp = (evt: React.MouseEvent) => {
-      onClick?.(evt as React.MouseEvent<HTMLTableCellElement>);
-      if (evt.button === 0) selectionDispatch({ type: SelectionActions.SELECT_MULTI_END });
-    };
-
-    const handleMouseClick = (evt: React.MouseEvent) => {
-      if (cell.isSelectable && evt.button === 0)
-        selectionDispatch({ type: SelectionActions.SELECT, payload: { cell, evt, announce } });
-    };
+    const { handleMouseDown, handleMouseOver, handleMouseUp, handleMouseClick } = getMouseHandlers(
+      cell,
+      announce,
+      onClick,
+      selectionDispatch
+    );
 
     const selectionStyling = getSelectionStyle(styling, cellSelectionState);
