+++ src/table/utils/handle-click.ts
@@ -1,7 +1,9 @@
 import { MouseEvent } from 'react';
 import { stardust } from '@nebula.js/stardust';
-import { Cell, TotalsPosition } from '../../types';
+import { Announce, Cell, TotalsPosition } from '../../types';
+import { SelectionActions } from './selections-utils';
 import { removeTabAndFocusCell, updateFocus, getCellElement } from './accessibility-utils';
+import { SelectionDispatch } from '../types';
 
 export const handleClickToFocusBody = (
   cell: Cell,
@@ -27,4 +29,35 @@ export const handleClickToFocusHead = (
 export const handleMouseDownLabelToFocusHeadCell = (evt: MouseEvent, rootElement: HTMLElement, columnIndex: number) => {
   evt.preventDefault();
   updateFocus({ focusType: 'focus', cell: getCellElement(rootElement, [0, columnIndex]) });
+};
+
+export const getMouseHandlers = (
+  cell: Cell,
+  announce: Announce,
+  onClick: React.MouseEventHandler<HTMLTableCellElement> | undefined,
+  selectionDispatch: SelectionDispatch
+) => {
+  const handleMouseDown = (evt: React.MouseEvent) => {
+    evt.preventDefault();
+    evt.stopPropagation();
+    cell.isSelectable && selectionDispatch({ type: SelectionActions.SELECT_MULTI_START, payload: { cell } });
+  };
+
+  const handleMouseOver = (evt: React.MouseEvent) => {
+    if (evt.button === 0 && evt.buttons === 1)
+      selectionDispatch({ type: SelectionActions.SELECT_MULTI_ADD, payload: { cell, evt, announce } });
+  };
+
+  const handleMouseUp = (evt: React.MouseEvent) => {
+    onClick?.(evt as React.MouseEvent<HTMLTableCellElement>);
+    (evt.target as HTMLElement).focus();
+    if (evt.button === 0) selectionDispatch({ type: SelectionActions.SELECT_MULTI_END });
+  };
+
+  const handleMouseClick = (evt: React.MouseEvent) => {
+    if (cell.isSelectable && evt.button === 0)
+      selectionDispatch({ type: SelectionActions.SELECT, payload: { cell, evt, announce } });
+  };
+
+  return { handleMouseDown, handleMouseOver, handleMouseUp, handleMouseClick };
 };
